requirements.txt
django-annoying==0.10.4
django-bower==5.2.0
django-debug-toolbar==1.9.1
django-phrase==1.0.6
djangorestframework==3.8.2
icalendar==4.0.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
virtualenv==16.0.0
